refactor: Eliminate some circular dependencies in AuxTreeCalculations
